Fixes issues in regrads to not placing fuel bottle
